# contracts/engine/openapi-lite.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LumaTorrent Engine API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}/v1",
      "variables": {
        "port": {
          "default": "43177"
        }
      }
    }
  ],
  "security": [
    {
      "EngineToken": []
    }
  ],
  "paths": {
...
    "/torrents/magnet": {
      "post": {
        "summary": "Add magnet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTorrentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Magnet accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddTorrentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/EngineError"
          }
        }
      }
    },
    "/torrents/file": {
      "post": {
        "summary": "Add torrent file",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTorrentFileRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Torrent file accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddTorrentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/EngineError"
          },
          "409": {
            "$ref": "#/components/responses/EngineError"
          }
        }
      }
    },
...
    "/torrents/{id}/remove": {
      "post": {
        "summary": "Remove torrent",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTorrentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Torrent removed safely",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveTorrentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/EngineError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EngineToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Luma-Engine-Token"
      }
    },
    "responses": {
      "EngineError": {
        "description": "Engine error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EngineError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "AddTorrentRequest": {
        "type": "object",
        "required": ["magnetUri", "savePath"],
        "properties": {
          "magnetUri": {
            "type": "string"
          },
          "savePath": {
            "type": "string"
          },
          "selectedFiles": {
            "type": ["array", "null"],
            "items": {
              "type": "string"
            }
          },
          "startPaused": {
            "type": "boolean"
          }
        }
      },
      "AddTorrentResponse": {
        "type": "object",
        "required": ["torrentId", "status"],
        "properties": {
          "torrentId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "checking",
              "metadata",
              "downloading",
              "paused",
              "completed",
              "seeding",
              "error"
            ]
          }
        }
      },
      "AddTorrentFileRequest": {
        "type": "object",
        "required": ["torrentFilePath", "savePath"],
        "properties": {
          "torrentFilePath": {
            "type": "string"
          },
          "savePath": {
            "type": "string"
          },
          "startPaused": {
            "type": "boolean"
          }
        }
      },
      "RemoveTorrentRequest": {
        "type": "object",
        "properties": {
          "deleteFiles": {
            "type": "boolean",
            "default": false
          },
          "useTrash": {
            "type": "boolean",
            "const": true,
            "default": true
          }
        }
      },
      "RemoveTorrentResponse": {
        "type": "object",
        "required": ["ok", "removedFromApp", "filesTrashed", "filesMissing"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "removedFromApp": {
            "type": "boolean"
          },
          "filesTrashed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filesMissing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EngineError": {
        "$ref": "./errors.schema.json"
      },
...
    }
  }
}